build - refactored and modified transaction service impl

+++ transaction-service/src/main/java/com/example/transactionservice/repository/TransactionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 
-    Optional<Transaction> findByRefNo(Integer refNo);
+    Optional<Transaction> findByRefNo(Long refNo);
 }
